First ASM changes ported from P+ 3.1.0
Adds Slippery, NewCommands and SubspaceSwap.

Not adding Camera for now, I need to see what it does in practice.

Not adding WallAttach and Hitbox for now, I need to verify their impact as well.

Moonwalk might be to avoid to preserve PM gameplay.

Effects might have a fix I'm interested in.